separate folder for httpv2 tests
