+++ tests/utils/testutils_docker.py
@@ -142,13 +142,13 @@ class WebHostDockerContainerBase(unittest.TestCase):
         install_libraries_cmd.extend(['--platform=manylinux2014_x86_64'])
         install_libraries_cmd.extend(configs.libraries)
         install_libraries_cmd.extend(['-t',
-                                        f'{script_path}/{_libraries_path}'])
+                                      f'{script_path}/{_libraries_path}'])
         install_libraries_cmd.extend(['--only-binary=:all:'])
 
         install_libraries_process = \
             subprocess.run(args=install_libraries_cmd,
-                            stdout=subprocess.PIPE,
-                            stderr=subprocess.PIPE)
+                           stdout=subprocess.PIPE,
+                           stderr=subprocess.PIPE)
 
         if install_libraries_process.returncode != 0:
             raise RuntimeError('Failed to install libraries')
